change int to unsigned long long (consider overflow)

+++ include/bank_api.hpp
@@ -8,14 +8,14 @@ public:
     BankAPI();
     void printAllCardInfo();
 
-    bool addCard(const std::string& card_id, const std::string& pin_number, int balance);
+    bool addCard(const std::string& card_id, const std::string& pin_number, unsigned long long balance);
 
     bool check_validity(const std::string& card_id); 
     bool verifyPIN(const std::string& card_id, const std::string& pin_number);
 
-    int seeBalance(const std::string& card_id);
-    void deposit(const std::string& card_id, int amount);
-    bool withdraw(const std::string& card_id, int amount);
+    unsigned long long seeBalance(const std::string& card_id);
+    void deposit(const std::string& card_id, unsigned long long amount);
+    bool withdraw(const std::string& card_id, unsigned long long amount);
 
 private:
     std::unordered_map<std::string, CardInfo> card_database;
